switch themes again

+++ bin/switch_php.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
 # ----------------------------------------------------------------------------
 # Switch between php versions including icu4c reinstall
-# @modified 05-Jun-2020
+# @modified 11-Jun-2020
 # ----------------------------------------------------------------------------
 
-PHP_FROM="7.1"
-PHP_TO="7.4"
-ICU4C_OLD="brew reinstall https://raw.githubusercontent.com/Homebrew/homebrew-core/896d1018c7a4906f2c3fa1386aaf283497db60a2/Formula/icu4c.rb"
-ICU4C_NEW="brew reinstall icu4c"
-PHP_DIR="/usr/local/opt"
+php_version="7.4"
+brew_prefix=$(brew --prefix)
+php_opt_dir="${brew_prefix}/opt"
+
+brew_array=("5.6","7.0","7.1","7.2","7.3","7.4")
+php_array=("php@5.6" "php@7.0" "php@7.1" "php@7.2" "php@7.3" "php@7.4")
 
 function usage()
 {
@@ -18,11 +19,9 @@ Usage:  $0 [OPTIONS]
 This script switches php versions
 - Stops running service
 - Relink php versions
-- Reinstall icu4c
 
 OPTIONS:
-  -f  From php (defaults to $PHP_FROM)
-  -t  To php (defaults to $PHP_TO)
+  -v  php version (default $PHP_VERSION)
   -h  Show this message
 
 EOF
@@ -35,13 +34,10 @@ function error()
 }
 
 PRINT=""
-while getopts "f:t:vh" OPTION; do
+while getopts "v:h" OPTION; do
   case $OPTION in
-    f)
-      PHP_FROM=$OPTARG
-      ;;
-    t)
-      PHP_TO=$OPTARG
+    v)
+      php_version=$OPTARG
       ;;
     h)
       usage
@@ -64,45 +60,60 @@ while getopts "f:t:vh" OPTION; do
   esac
 done
 
-LAST_DIR=$(pwd)
+php_version="php@$php_version"
+printf "Processing $php_version\n"
 
-echo "brew services stop php/php@$PHP_FROM"
-brew services stop php
-brew services stop php@$PHP_FROM
+last_dir=$(pwd)
+cd $php_opt_dir
 
-echo ""
-echo "brew unlink php@$PHP_FROM && brew link --force --overwrite php@$PHP_TO"
-brew unlink php@$PHP_FROM && brew link --force --overwrite php@$PHP_TO
+# What versions of php are installed via brew
+printf " - check installed php versions: "
+for i in ${php_array[*]}; do
+    if [[ -n "$(brew ls --versions "$i")" ]]; then
+        php_installed_array+=("$i")
+    fi
+done
+echo ${php_installed_array[*]}
 
-echo "ln -s php@$PHP_TO php"
-cd $PHP_DIR
-rm php
-ln -s php@$PHP_TO php
-
-echo ""
-if [ "$PHP_TO" == '5.6' ]
-then
-  echo $ICU4C_OLD
-  eval $ICU4C_OLD
-elif [ "$PHP_TO" == '7.1' ]
-then
-  echo $ICU4C_OLD
-  eval $ICU4C_OLD
-elif [ "$PHP_TO" == '7.2' ]
-then
-  echo $ICU4C_NEW
-  eval $ICU4C_NEW
-elif [ "$PHP_TO" == '7.4' ]
-then
-  echo $ICU4C_NEW
-  eval $ICU4C_NEW
+# Stop php service
+printf " - stopping php services: "
+php_running=`brew services | grep php | grep started | cut -f1 -d' '`
+if [[ -n $php_running ]]; then
+    echo "$php_running"
+    brew services stop $php_running
+else
+    echo "none"
 fi
 
-echo ""
-echo "brew services start php@$PHP_TO"
-brew services start php@$PHP_TO
+# Check that the requested version is supported
+printf " - checking version supported: "
+if [[ " ${php_array[*]} " == *"$php_version"* ]]; then
+    # Check that the requested version is installed
+    if [[ " ${php_installed_array[*]} " == *"$php_version"* ]]; then
+        printf "OK\n"
+        echo " - relink to brew $php_version"
+        for i in ${php_installed_array[@]}; do
+            brew unlink $i
+        done
+        brew link --force "$php_version"
+
+        echo " - relink php symlink"
+        symlink=`ls -l php_version | cut -f14 -d' '`
+        rm php
+        ln -s $symlink php
+
+        echo " - restart php service"
+        brew services start $php_version
+    else
+        echo "Sorry, but $php_version is not installed via brew. Install by running: brew install $php_version"
+    fi
+else
+    echo "Unknown version of PHP. PHP Switcher can only handle arguments of:" ${brew_array[@]}
+fi
 
 ps auwx | grep php
 php -v
+echo " - finished"
 
-cd $LAST_DIR
+cd $last_dir
+exit
